# backend/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Football Website API",
    "version": "1.0.0",
    "description": "Public read API (v1) for news, matches, teams, players, competitions and transfers. All list responses are paginated. Errors use a stable { error: { code, message } } envelope."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [],
  "paths": {
...
    "/teams": {
      "get": {
        "summary": "List teams",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated teams"
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
    "/teams/{slug}": {
      "get": {
        "summary": "Get team by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/teams/{slug}/details": {
      "get": {
        "summary": "Team page (competitions, matches, articles, squad)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team details"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/transfers": {
      "get": {
        "summary": "List public transfers (announced/completed only)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "announced",
                "completed"
              ]
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "permanent",
                "loan",
                "loan_return",
                "free_transfer"
              ]
            }
          },
          {
            "name": "player",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromTeam",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toTeam",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "window",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "announcedFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "announcedTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "effectiveFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "effectiveTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated transfers"
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
    "/transfers/{id}": {
      "get": {
        "summary": "Get public transfer by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer"
          },
          "400": {
            "description": "Invalid id"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}